refactor: separate mod types
commit-id:09a283bc
