+++ extensions/scarb-doc/src/docs_generation/markdown/context.rs
@@ -1,6 +1,6 @@
 use crate::docs_generation::markdown::SUMMARY_FILENAME;
 use crate::docs_generation::markdown::traits::WithPath;
-use crate::types::Crate;
+use crate::types::crate_type::Crate;
 use cairo_lang_defs::ids::{ImplItemId, LookupItemId, TraitItemId};
 use cairo_lang_doc::documentable_item::DocumentableItemId;
 use cairo_lang_doc::parser::CommentLinkToken;
